convo-genius-ai-studio/backend/app/api/code_generation.py
```python
"""
ConvoGenius AI Studio: Crafting Conversations, Code, and Dreams
Code Generation API
"""

import os
import sys

sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))

from models.code_synthesis_model import CodeSynthesisModel
from models.nlp_model import NLPModel
from utils.web_scraping import scrape_website
from utils.computer_vision import process_screenshot

def generate_code(intent):
    """
    Generate code based on user intent
    """
    if intent == "web_scraping":
        return scrape_website()
    elif intent == "screen_interaction":
        return process_screenshot()
    elif intent == "app_development":
        return CodeSynthesisModel.generate_app_code()
    else:
        return NLPModel.generate_code(intent)
```