convo-genius-ai-studio/backend/tests/test_code_generation.py
```python
import unittest
from app.api.code_generation import generate_code

class TestCodeGeneration(unittest.TestCase):
    def test_generate_code(self):
        # Test case 1
        intent = "create a React component"
        code = generate_code(intent)
        expected_code = """
import React from 'react';

const MyComponent = () => {
  return (
    <div>
      {/* Your component code here */}
    </div>
  );
};

export default MyComponent;
"""
        self.assertEqual(code.strip(), expected_code.strip())

        # Test case 2
        intent = "fetch data from an API"
        code = generate_code(intent)
        expected_code = """
import axios from 'axios';

const fetchData = async () => {
  try {
    const response = await axios.get('https://api.example.com/data');
    const data = response.data;
    // Process the data here
  } catch (error) {
    console.error('Error fetching data:', error);
  }
};

fetchData();
"""
        self.assertEqual(code.strip(), expected_code.strip())

        # Test case 3
        intent = "perform screen interaction"
        code = generate_code(intent)
        expected_code = """
import cv2
import pyautogui

def perform_screen_interaction():
  # Capture screen
  screenshot = pyautogui.screenshot()
  screenshot.save('screenshot.png')

  # Process the screenshot using OpenCV
  image = cv2.imread('screenshot.png')
  # Perform screen interaction tasks here

perform_screen_interaction()
"""
        self.assertEqual(code.strip(), expected_code.strip())

if __name__ == '__main__':
    unittest.main()
```